+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 #### 🦟 Bugs Fixed
 
 -   Amethyst Golems now clean up after themselves [(\#97)](https://github.com/EnigmaticaModpacks/Enigmatica10/issues/97)
+-   Aquaculture fish are now tagged as fish [(\#97)](https://github.com/EnigmaticaModpacks/Enigmatica10/issues/97)
 
 ### Enigmatica 10 1.9.0